+++ src/main/java/flab/nutridiary/diary/service/AddDiaryRecordService.java
@@ -5,8 +5,8 @@ import flab.nutridiary.diary.domain.Diary;
 import flab.nutridiary.diary.domain.DiaryRecord;
 import flab.nutridiary.diary.domain.NutritionCalculator;
 import flab.nutridiary.diary.domain.ProductIntakeInfo;
-import flab.nutridiary.diary.dto.AddDiaryRecordRequest;
-import flab.nutridiary.diary.dto.DiarySavedResponse;
+import flab.nutridiary.diary.dto.request.AddDiaryRecordRequest;
+import flab.nutridiary.diary.dto.response.DiarySavedResponse;
 import flab.nutridiary.diary.repository.DiaryRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
