+++ src/test/java/flab/nutridiary/diary/controller/DiaryControllerTest.java
@@ -6,8 +6,8 @@ import flab.nutridiary.product.domain.ServingUnit;
 import flab.nutridiary.diary.domain.Diary;
 import flab.nutridiary.diary.domain.DiaryRecord;
 import flab.nutridiary.diary.domain.MealType;
-import flab.nutridiary.diary.dto.AddDiaryRecordRequest;
-import flab.nutridiary.diary.dto.DiaryRegisterRequest;
+import flab.nutridiary.diary.dto.request.AddDiaryRecordRequest;
+import flab.nutridiary.diary.dto.request.DiaryRegisterRequest;
 import flab.nutridiary.diary.repository.DiaryRepository;
 import flab.nutridiary.product.domain.NutritionFacts;
 import flab.nutridiary.product.domain.Product;
