+++ src/test/java/flab/nutridiary/diary/service/DiaryRegisterServiceTest.java
@@ -6,9 +6,9 @@ import flab.nutridiary.product.domain.ServingUnit;
 import flab.nutridiary.diary.domain.Diary;
 import flab.nutridiary.diary.domain.DiaryRecord;
 import flab.nutridiary.diary.domain.MealType;
-import flab.nutridiary.diary.dto.AddDiaryRecordRequest;
-import flab.nutridiary.diary.dto.DiaryRegisterRequest;
-import flab.nutridiary.diary.dto.DiarySavedResponse;
+import flab.nutridiary.diary.dto.request.AddDiaryRecordRequest;
+import flab.nutridiary.diary.dto.request.DiaryRegisterRequest;
+import flab.nutridiary.diary.dto.response.DiarySavedResponse;
 import flab.nutridiary.diary.repository.DiaryRepository;
 import flab.nutridiary.product.domain.NutritionFacts;
 import flab.nutridiary.product.domain.Product;
