+++ src/test/java/flab/nutridiary/diary/service/ProductIntakeInfoMapperTest.java
@@ -2,7 +2,7 @@ package flab.nutridiary.diary.service;
 
 import flab.nutridiary.diary.domain.MealType;
 import flab.nutridiary.diary.domain.ProductIntakeInfo;
-import flab.nutridiary.diary.dto.DiaryRegisterRequest;
+import flab.nutridiary.diary.dto.request.DiaryRegisterRequest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
